+++ src/main/java/api/impl/ClientImpl.java
@@ -1,6 +1,8 @@
 package api.impl;
 
+import api.dao.AccountRepository;
 import api.dao.ClientRepository;
+import api.dao.entity.Account;
 import api.dao.entity.Client;
 import api.data.ClientSearchData;
 import jakarta.enterprise.context.ApplicationScoped;
@@ -16,26 +18,39 @@ public class ClientImpl {
     @Inject
     ClientRepository clientRepository;
 
-    public List<Client> listClient(){
+    @Inject
+    AccountRepository accountRepository;
+
+    public List<Client> listClient() {
         return clientRepository.listAll();
     }
-    public Client save(Client client){
+
+    public Client save(Client client) {
         clientRepository.persist(client);
         return client;
     }
-    public boolean delete(Client client){
-        return clientRepository.deleteById(client.getId());
+
+    public boolean delete(Client client) {
+        Account account = accountRepository.find("select o from Account o where o.client.id =" + client.getId()).firstResult();
+        if (account == null) {
+
+            clientRepository.deleteById(client.getId());
+            return true;
+        }
+        return false;
     }
-    public Client search(ClientSearchData searchData){
+
+    public Client search(ClientSearchData searchData) {
         Map<String, Object> params = new HashMap<>();
         params.put("documentType", searchData.getDocumentType());
         params.put("documentNumber", searchData.getDocumentNumber());
 
         return clientRepository.find(
-                "SELECT o FROM  Client o WHERE o.documentType=:documentType AND o.documentNumber=:documentNumber",params
+                "SELECT o FROM  Client o WHERE o.documentType=:documentType AND o.documentNumber=:documentNumber", params
         ).firstResult();
     }
-    public Client update(Client client){
+
+    public Client update(Client client) {
         Client entity = clientRepository.findById(client.getId());
         entity.setName(client.getName());
         entity.setBirthday(client.getBirthday());
@@ -49,7 +64,4 @@ public class ClientImpl {
         return client;
     }
 
-    public Client findById(Long clientId){
-        return clientRepository.findById(clientId);
-    }
 }
